Stop document from scrolling behind lightbox on mobile devices
The HTML element now needs 4 CSS properties set instead of 1. Instead
of assuming that none are set and that we can override them at any time
we wish, we simply add and remove a class to set and 3 properties, and
set the other with JS as it is dynamic, and based on scroll position.