Optimize landscape view for small screens and more
Reorganize CSS: nesting is unnecessary and even harmful for when users
will want to override CSS rules with their own style. Add breakpoints
on height as well as width to take care of landscape views on small
screens. On small screens, have the controls and title overlay the
photo. Normalize padding of controls.